Move the methods for schema dumping into `{mysql,postgresql}/schema_dumper.rb`
Current master branch includes many schema dumping improvements.
It extract these features to the appropriate files.
